+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - binary compatibility will now be verified and held up on every release.
 
+## `[0.15.0]` - 2021-XX-XX
+
+- Refactor `Controller.state` from `Flow<State>` to `StateFlow<State>`.
+- Deprecate `Controller.currentState`.
+- Internal `BroadcastChannel` implementations changed to `SharedFlow`.
+
 ## `[0.14.0]` - 2021-10-17
 
 - Update kotlinx.coroutines to `1.5.2`.